fix: keep the whole URL when redirecting after log-in
When we navigate to any page being logged-out we are redirecting to login page. The same time we should keep the FULL URL which we have before so after login we can redirect back to the same page we were trying to navigate.

Before this fix we only keep a part of URL so sometimes we redirect back not to the exact page we were going to intially.